backend/requirements.txt
flask
flask-cors
python-dotenv
openai
google-generativeai
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.1
requests==2.31.0
werkzeug==3.0.1